+++ config/routes.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace ICanBoogie\Responder;
-
-use ICanBoogie\Binding\Routing\ConfigBuilder;
-
-return fn(ConfigBuilder $config) => $config
-    ->get('/api/ping', 'api:ping');